+++ R/create_test_sleeplog_csv.R
@@ -1,18 +1,23 @@
 create_test_sleeplog_csv = function(Nnights = 7,
                                     storagelocation = c(),
-                                    advanced = FALSE,
-                                    sep = ",") {
+                                    advanced = FALSE, sep = ",",
+                                    missingOnset = FALSE, missingWake = FALSE) {
   
   # function to create a test sleeplog file needed for testing GGIR
   # Nnights is the number of nights the sleeplog has
   if (length(storagelocation) == 0) storagelocation = getwd()
   if (Nnights == 0) Nnights = 1
   if (advanced == FALSE) {
-    sleeplog = as.data.frame(t(c("123",rep(c("23:00:00","07:00:00"),Nnights))), stringsAsFactors = TRUE)
+    # create times where the second vlaue increase by one every day
+    times = rep(c("23:00:","07:00:"), Nnights)
+    times = paste0(times, paste0("0", rep(1:Nnights, each = 2)))
+    sleeplog = as.data.frame(t(c("123", times)), stringsAsFactors = TRUE)
     if (ncol(sleeplog) >= 5) {
       # let's add one night where the person slept untill the afternoon
       sleeplog[, 4:5] = c("03:00:00", "17:00:00") 
     }
+    if (missingOnset == TRUE) sleeplog[, 2] = ""
+    if (missingWake == TRUE) sleeplog[, 3] = ""
     colnames(sleeplog) = c("ID",rep(c("onset","wakeup"),Nnights))
     write.table(
       sleeplog,
@@ -23,7 +28,10 @@ create_test_sleeplog_csv = function(Nnights = 7,
       fileEncoding = "UTF-8"
     )
   } else {
-    times = c("07:00:00", "23:00:00", "13:00:00", "13:30:00", "15:00:00", "15:40:00",  "17:00:00", "17:15:00")
+    # Note for advanced log we will need one extra date, 
+    # because the advanced log is calendar day oriented
+    times = c("07:00:00", "23:00:00", "13:00:00", "13:30:00",
+              "15:00:00", "15:40:00",  "17:00:00", "17:15:00")
     dates = seq(as.Date("2016/06/25"), as.Date("2016/06/25") + Nnights, by = 1)
     Cnames = c(
       "_wakeup",
@@ -44,6 +52,10 @@ create_test_sleeplog_csv = function(Nnights = 7,
       } else {
         times2 = gsub(pattern = ":00:00", replacement = paste0(":00:", ifelse(di < 59, yes = di, no = "00")),
                                                                x = times)
+      }
+      if (di == 1) {
+        if (missingOnset == TRUE) times2[1] = ""
+        if (missingWake == TRUE) times2[2] = ""
       }
       log = c(log, as.character(dates[di]), times2)
       names = c(names, c(paste0("D",di,"_date"),paste0("D",di, Cnames)))